supabase-mcp-client/requirements.txt
anthropic==0.37.1
mcp==1.3.3
python-dotenv==1.0.1
